Unicode handling for harvested documents
All remote documents stored on the database are converted to unicode.
This is done via requests (bumped to the latest version) or lxml in the
CSW case (see #9ea721e on ckanext-spatial).

The XML declaration is also updated to reflect that the content is
encoded in UTF-8 (see #866fd69 on ckanext-harvest).