Add links to different sourcxe metadata files
Adds links on the dataset page to the relevant harvest object contents,
both raw and formatted as html. The HTML view does not work because the
provided XSLT does not support neither ISO nor FGDC documents.

If the original data was in FGDC, an additional set of links are
displayed pointing to it.