[#43] Map widget for drawing a spatial filter
On the filter section of the dataset search page there is a now a map
widget that allows users to define an area of interest to limit search
results, in the same way that the facets do.

The bbox and extent are passed as parameters so they can be reproduced
on the resulting page. The map is powered by Leaflet.