[#43] Improve search when looking up a location
The map widget will now automatically center to a provided bbox on
startup. This allows that when looking up a location on the autocomplete
field, once you pick a result there is no need to draw the box in the map
before submitting the form, just storing its value in the ext_bbox
field. This has the added benefit of decoupling the two widgets in case
they want to be used separately, as they are now independent (although
related via the ext_bbox field).